Improved Code-3
